fixed total price not showing in receipts (has bug when more details are clicked)

+++ app/src/main/java/com/example/washuplaundry/ReceiptDataRow.kt
@@ -1,11 +1,9 @@
 package com.example.washuplaundry
 
-import java.math.BigDecimal
-
 data class ReceiptDataRow(
     val joNumber: String,
     val timestamp: String?,
     val orderItems: List<OrderData>,
-    val totalPrice: BigDecimal,
+    val totalPrice: Double,
     var isExpanded: Boolean = false
 )
